[XEB] Split into three files
