+++ cirq/experiments/fidelity_estimation_test.py
@@ -13,27 +13,12 @@
 # limitations under the License.
 import itertools
 import math
-import multiprocessing
-import time
-from typing import Dict, Any, cast, Optional
-from typing import Sequence, Iterable
+from typing import Sequence
 
-import networkx as nx
 import numpy as np
-import pandas as pd
 import pytest
 
 import cirq
-import cirq.experiments.random_quantum_circuit_generation as rqcg
-from cirq.experiments.fidelity_estimation import (
-    SQRT_ISWAP,
-    sample_2q_xeb_circuits,
-    simulate_2q_xeb_circuits,
-    benchmark_2q_xeb_fidelities,
-    parameterize_phased_fsim_circuit,
-    characterize_phased_fsim_parameters_with_xeb,
-    SqrtISwapXEBOptions,
-)
 
 
 def sample_noisy_bitstrings(
@@ -254,423 +239,3 @@ def test_least_squares_xeb_fidelity_from_probabilities():
     np.testing.assert_allclose(np.sum(np.array(r_log_math) ** 2), 0.0, atol=1e-2)
 
     np.random.set_state(prng_state)
-
-
-def test_sample_2q_xeb_circuits():
-    q0 = cirq.NamedQubit('a')
-    q1 = cirq.NamedQubit('b')
-    circuits = [
-        rqcg.random_rotations_between_two_qubit_circuit(
-            q0,
-            q1,
-            depth=20,
-            two_qubit_op_factory=lambda a, b, _: SQRT_ISWAP(a, b),
-        )
-        for _ in range(2)
-    ]
-    cycle_depths = np.arange(3, 20, 6)
-
-    df = sample_2q_xeb_circuits(
-        sampler=cirq.Simulator(),
-        circuits=circuits,
-        cycle_depths=cycle_depths,
-    )
-    assert len(df) == len(cycle_depths) * len(circuits)
-    for (circuit_i, cycle_depth), row in df.iterrows():
-        assert 0 <= circuit_i < len(circuits)
-        assert cycle_depth in cycle_depths
-        assert len(row['sampled_probs']) == 4
-        assert np.isclose(np.sum(row['sampled_probs']), 1)
-
-
-def test_sample_2q_xeb_circuits_error():
-    qubits = cirq.LineQubit.range(3)
-    circuits = [cirq.testing.random_circuit(qubits, n_moments=5, op_density=0.8, random_state=52)]
-    cycle_depths = np.arange(3, 50, 9)
-    with pytest.raises(ValueError):  # three qubit circuits
-        _ = sample_2q_xeb_circuits(
-            sampler=cirq.Simulator(),
-            circuits=circuits,
-            cycle_depths=cycle_depths,
-        )
-
-
-def test_sample_2q_xeb_circuits_no_progress(capsys):
-    qubits = cirq.LineQubit.range(2)
-    circuits = [cirq.testing.random_circuit(qubits, n_moments=7, op_density=0.8, random_state=52)]
-    cycle_depths = np.arange(3, 4)
-    _ = sample_2q_xeb_circuits(
-        sampler=cirq.Simulator(),
-        circuits=circuits,
-        cycle_depths=cycle_depths,
-        progress_bar=None,
-    )
-    captured = capsys.readouterr()
-    assert captured.out == ''
-    assert captured.err == ''
-
-
-def _gridqubits_to_graph_device(qubits: Iterable[cirq.GridQubit]):
-    # cirq contrib: routing.gridqubits_to_graph_device
-    def _manhattan_distance(qubit1: cirq.GridQubit, qubit2: cirq.GridQubit) -> int:
-        return abs(qubit1.row - qubit2.row) + abs(qubit1.col - qubit2.col)
-
-    return nx.Graph(
-        pair for pair in itertools.combinations(qubits, 2) if _manhattan_distance(*pair) == 1
-    )
-
-
-def test_sample_2q_parallel_xeb_circuits():
-    circuits = rqcg.generate_library_of_2q_circuits(
-        n_library_circuits=5, two_qubit_gate=cirq.ISWAP ** 0.5, max_cycle_depth=10
-    )
-    cycle_depths = [10]
-    graph = _gridqubits_to_graph_device(cirq.GridQubit.rect(3, 2))
-    combs = rqcg.get_random_combinations_for_device(
-        n_library_circuits=len(circuits),
-        n_combinations=5,
-        device_graph=graph,
-        random_state=10,
-    )
-
-    df = sample_2q_xeb_circuits(
-        sampler=cirq.Simulator(),
-        circuits=circuits,
-        cycle_depths=cycle_depths,
-        combinations_by_layer=combs,
-    )
-    n_pairs = sum(len(c.pairs) for c in combs)
-    assert len(df) == len(cycle_depths) * len(circuits) * n_pairs
-    for (circuit_i, cycle_depth), row in df.iterrows():
-        assert 0 <= circuit_i < len(circuits)
-        assert cycle_depth in cycle_depths
-        assert len(row['sampled_probs']) == 4
-        assert np.isclose(np.sum(row['sampled_probs']), 1)
-        assert 0 <= row['layer_i'] < 4
-        assert 0 <= row['pair_i'] < 2  # in 3x2 graph, there's a max of 2 pairs per layer
-    assert len(df['pair_name'].unique()) == 7  # seven pairs in 3x2 graph
-
-
-def test_sample_2q_parallel_xeb_circuits_bad_circuit_library():
-    circuits = rqcg.generate_library_of_2q_circuits(
-        n_library_circuits=5, two_qubit_gate=cirq.ISWAP ** 0.5, max_cycle_depth=10
-    )
-    cycle_depths = [10]
-    graph = _gridqubits_to_graph_device(cirq.GridQubit.rect(3, 2))
-    combs = rqcg.get_random_combinations_for_device(
-        n_library_circuits=len(circuits) + 100,  # !!! should cause invlaid input
-        n_combinations=5,
-        device_graph=graph,
-        random_state=10,
-    )
-
-    with pytest.raises(ValueError, match='.*invalid indices.*'):
-        _ = sample_2q_xeb_circuits(
-            sampler=cirq.Simulator(),
-            circuits=circuits,
-            cycle_depths=cycle_depths,
-            combinations_by_layer=combs,
-        )
-
-
-def test_sample_2q_parallel_xeb_circuits_error_bad_qubits():
-    circuits = rqcg.generate_library_of_2q_circuits(
-        n_library_circuits=5,
-        two_qubit_gate=cirq.ISWAP ** 0.5,
-        max_cycle_depth=10,
-        q0=cirq.GridQubit(0, 0),
-        q1=cirq.GridQubit(1, 1),
-    )
-    cycle_depths = [10]
-    graph = _gridqubits_to_graph_device(cirq.GridQubit.rect(3, 2))
-    combs = rqcg.get_random_combinations_for_device(
-        n_library_circuits=len(circuits),
-        n_combinations=5,
-        device_graph=graph,
-        random_state=10,
-    )
-
-    with pytest.raises(ValueError, match=r'.*each operating on LineQubit\(0\) and LineQubit\(1\)'):
-        _ = sample_2q_xeb_circuits(
-            sampler=cirq.Simulator(),
-            circuits=circuits,
-            cycle_depths=cycle_depths,
-            combinations_by_layer=combs,
-        )
-
-
-def test_simulate_2q_xeb_circuits():
-    q0, q1 = cirq.LineQubit.range(2)
-    circuits = [
-        rqcg.random_rotations_between_two_qubit_circuit(
-            q0,
-            q1,
-            depth=50,
-            two_qubit_op_factory=lambda a, b, _: SQRT_ISWAP(a, b),
-        )
-        for _ in range(2)
-    ]
-    cycle_depths = np.arange(3, 50, 9)
-
-    df = simulate_2q_xeb_circuits(
-        circuits=circuits,
-        cycle_depths=cycle_depths,
-    )
-    assert len(df) == len(cycle_depths) * len(circuits)
-    for (circuit_i, cycle_depth), row in df.iterrows():
-        assert 0 <= circuit_i < len(circuits)
-        assert cycle_depth in cycle_depths
-        assert len(row['pure_probs']) == 4
-        assert np.isclose(np.sum(row['pure_probs']), 1)
-
-    with multiprocessing.Pool() as pool:
-        df2 = simulate_2q_xeb_circuits(circuits, cycle_depths, pool=pool)
-
-    pd.testing.assert_frame_equal(df, df2)
-
-
-def test_simulate_circuit_length_validation():
-    q0, q1 = cirq.LineQubit.range(2)
-    circuits = [
-        rqcg.random_rotations_between_two_qubit_circuit(
-            q0,
-            q1,
-            depth=10,  # not long enough!
-            two_qubit_op_factory=lambda a, b, _: SQRT_ISWAP(a, b),
-        )
-        for _ in range(2)
-    ]
-    cycle_depths = np.arange(3, 50, 9)
-    with pytest.raises(ValueError, match='.*not long enough.*'):
-        _ = simulate_2q_xeb_circuits(
-            circuits=circuits,
-            cycle_depths=cycle_depths,
-        )
-
-
-def test_benchmark_2q_xeb_fidelities():
-    q0, q1 = cirq.LineQubit.range(2)
-    circuits = [
-        rqcg.random_rotations_between_two_qubit_circuit(
-            q0, q1, depth=50, two_qubit_op_factory=lambda a, b, _: SQRT_ISWAP(a, b), seed=52
-        )
-        for _ in range(2)
-    ]
-    cycle_depths = np.arange(3, 50, 9)
-
-    sampled_df = sample_2q_xeb_circuits(
-        sampler=cirq.Simulator(seed=53), circuits=circuits, cycle_depths=cycle_depths
-    )
-    fid_df = benchmark_2q_xeb_fidelities(sampled_df, circuits, cycle_depths)
-    assert len(fid_df) == len(cycle_depths)
-    for _, row in fid_df.iterrows():
-        assert row['cycle_depth'] in cycle_depths
-        assert row['fidelity'] > 0.98
-
-
-def test_benchmark_2q_xeb_fidelities_parallel():
-    circuits = rqcg.generate_library_of_2q_circuits(
-        n_library_circuits=5, two_qubit_gate=cirq.ISWAP ** 0.5, max_cycle_depth=10
-    )
-    cycle_depths = [10]
-    graph = _gridqubits_to_graph_device(cirq.GridQubit.rect(2, 2))
-    combs = rqcg.get_random_combinations_for_device(
-        n_library_circuits=len(circuits),
-        n_combinations=2,
-        device_graph=graph,
-        random_state=10,
-    )
-
-    sampled_df = sample_2q_xeb_circuits(
-        sampler=cirq.Simulator(),
-        circuits=circuits,
-        cycle_depths=cycle_depths,
-        combinations_by_layer=combs,
-    )
-    fid_df = benchmark_2q_xeb_fidelities(sampled_df, circuits, cycle_depths)
-    n_pairs = sum(len(c.pairs) for c in combs)
-    assert len(fid_df) == len(cycle_depths) * n_pairs
-
-
-def test_parameterize_phased_fsim_circuit():
-    q0, q1 = cirq.LineQubit.range(2)
-    circuit = rqcg.random_rotations_between_two_qubit_circuit(
-        q0, q1, depth=3, two_qubit_op_factory=lambda a, b, _: SQRT_ISWAP(a, b), seed=52
-    )
-
-    p_circuit = parameterize_phased_fsim_circuit(circuit, SqrtISwapXEBOptions())
-    cirq.testing.assert_has_diagram(
-        p_circuit,
-        """\
-0                                    1
-│                                    │
-Y^0.5                                X^0.5
-│                                    │
-PhFSim(theta, zeta, chi, gamma, phi)─PhFSim(theta, zeta, chi, gamma, phi)
-│                                    │
-PhX(0.25)^0.5                        Y^0.5
-│                                    │
-PhFSim(theta, zeta, chi, gamma, phi)─PhFSim(theta, zeta, chi, gamma, phi)
-│                                    │
-Y^0.5                                X^0.5
-│                                    │
-PhFSim(theta, zeta, chi, gamma, phi)─PhFSim(theta, zeta, chi, gamma, phi)
-│                                    │
-X^0.5                                PhX(0.25)^0.5
-│                                    │
-    """,
-        transpose=True,
-    )
-
-
-def test_get_initial_simplex():
-    options = SqrtISwapXEBOptions()
-    simplex, names = options.get_initial_simplex_and_names()
-    assert names == ['theta', 'zeta', 'chi', 'gamma', 'phi']
-    assert len(simplex) == len(names) + 1
-    assert simplex.shape[1] == len(names)
-
-
-def test_characterize_phased_fsim_parameters_with_xeb():
-    q0, q1 = cirq.LineQubit.range(2)
-    rs = np.random.RandomState(52)
-    circuits = [
-        rqcg.random_rotations_between_two_qubit_circuit(
-            q0,
-            q1,
-            depth=20,
-            two_qubit_op_factory=lambda a, b, _: SQRT_ISWAP(a, b),
-            seed=rs,
-        )
-        for _ in range(2)
-    ]
-    cycle_depths = np.arange(3, 20, 6)
-    sampled_df = sample_2q_xeb_circuits(
-        sampler=cirq.Simulator(seed=rs),
-        circuits=circuits,
-        cycle_depths=cycle_depths,
-        progress_bar=None,
-    )
-    # only optimize theta so it goes faster.
-    options = SqrtISwapXEBOptions(
-        characterize_theta=True,
-        characterize_gamma=False,
-        characterize_chi=False,
-        characterize_zeta=False,
-        characterize_phi=False,
-    )
-    p_circuits = [parameterize_phased_fsim_circuit(circuit, options) for circuit in circuits]
-    with multiprocessing.Pool() as pool:
-        result = characterize_phased_fsim_parameters_with_xeb(
-            sampled_df=sampled_df,
-            parameterized_circuits=p_circuits,
-            cycle_depths=cycle_depths,
-            phased_fsim_options=options,
-            # speed up with looser tolerances:
-            fatol=1e-2,
-            xatol=1e-2,
-            pool=pool,
-        )
-    assert np.abs(result.x[0] + np.pi / 4) < 0.1
-    assert np.abs(result.fun) < 0.1  # noiseless simulator
-
-
-def _ref_simulate_2q_xeb_circuit(task: Dict[str, Any]):
-    """Helper function for simulating a given (circuit, cycle_depth)."""
-    circuit_i = task['circuit_i']
-    cycle_depth = task['cycle_depth']
-    circuit = task['circuit']
-    param_resolver = task['param_resolver']
-
-    circuit_depth = cycle_depth * 2 + 1
-    assert circuit_depth <= len(circuit)
-    tcircuit = circuit[:circuit_depth]
-    tcircuit = cirq.resolve_parameters_once(tcircuit, param_resolver=param_resolver)
-
-    pure_sim = cirq.Simulator()
-    psi = cast(cirq.StateVectorTrialResult, pure_sim.simulate(tcircuit))
-    psi = psi.final_state_vector
-    pure_probs = np.abs(psi) ** 2
-
-    return {
-        'circuit_i': circuit_i,
-        'cycle_depth': cycle_depth,
-        'pure_probs': pure_probs,
-    }
-
-
-def _ref_simulate_2q_xeb_circuits(
-    circuits: Sequence['cirq.Circuit'],
-    cycle_depths: Sequence[int],
-    param_resolver: 'cirq.ParamResolverOrSimilarType' = None,
-    pool: Optional['multiprocessing.pool.Pool'] = None,
-):
-    """Reference implementation for `simulate_2q_xeb_circuits` that
-    does each circuit independently instead of using intermediate states.
-
-    You can also try editing the helper function to use QSimSimulator() for
-    benchmarking. This simulator does not support intermediate states, so
-    you can't use it with the new functionality.
-    https://github.com/quantumlib/qsim/issues/101
-    """
-    tasks = []
-    for cycle_depth in cycle_depths:
-        for circuit_i, circuit in enumerate(circuits):
-            tasks += [
-                {
-                    'circuit_i': circuit_i,
-                    'cycle_depth': cycle_depth,
-                    'circuit': circuit,
-                    'param_resolver': param_resolver,
-                }
-            ]
-
-    if pool is not None:
-        records = pool.map(_ref_simulate_2q_xeb_circuit, tasks, chunksize=4)
-    else:
-        records = [_ref_simulate_2q_xeb_circuit(record) for record in tasks]
-
-    return pd.DataFrame(records).set_index(['circuit_i', 'cycle_depth']).sort_index()
-
-
-@pytest.mark.parametrize('multiprocess', (True, False))
-def test_incremental_simulate(multiprocess):
-    q0, q1 = cirq.LineQubit.range(2)
-    circuits = [
-        rqcg.random_rotations_between_two_qubit_circuit(
-            q0,
-            q1,
-            depth=100,
-            two_qubit_op_factory=lambda a, b, _: SQRT_ISWAP(a, b),
-        )
-        for _ in range(20)
-    ]
-    cycle_depths = np.arange(3, 100, 9)
-
-    if multiprocess:
-        pool = multiprocessing.Pool()
-    else:
-        pool = None
-
-    start = time.perf_counter()
-    df_ref = _ref_simulate_2q_xeb_circuits(
-        circuits=circuits,
-        cycle_depths=cycle_depths,
-        pool=pool,
-    )
-    end1 = time.perf_counter()
-
-    df = simulate_2q_xeb_circuits(circuits=circuits, cycle_depths=cycle_depths, pool=pool)
-    end2 = time.perf_counter()
-    if pool is not None:
-        pool.terminate()
-    print("\nnew:", end2 - end1, "old:", end1 - start)
-
-    pd.testing.assert_frame_equal(df_ref, df)
-
-    # Use below for approximate equality, if e.g. you're using qsim:
-    # assert len(df_ref) == len(df)
-    # assert df_ref.columns == df.columns
-    # for (i1, row1), (i2, row2) in zip(df_ref.iterrows(), df.iterrows()):
-    #     assert i1 == i2
-    #     np.testing.assert_allclose(row1['pure_probs'], row2['pure_probs'], atol=5e-5)
